Rename & Move: 컴포넌트 폴더 정리
