+++ src/components/pages/activities-detail.tsx/activities-description.tsx
