+++ src/components/pages/home/search-section.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image";
-import Button from "../../ui/button";
+import Button from "../../common/ui/button";
 
 export default function SearchSection() {
   return (
